brought out types and mock data to colocated file
